fix the lib64 path???
git-svn-id: http://172.26.0.2/svn/sevendeset/7d7/trunk@159 1d0ef2df-3fe9-40e7-8648-ba97643a2676

+++ bootstrap/BootStrap.py
@@ -116,9 +116,10 @@ Check documentation of the 'linux32' or the 'util-linux' Debian/Ubuntu packages
             os.makedirs(self._outputPath,0755)
         if not os.path.exists(self._prefix):
             os.makedirs(self._prefix,0755)
-            if self._arch == 'x86_64':
+        if not os.path.exists(self._prefix+"/lib64") and self._arch == 'x86_64':
+            if not os.path.exists(self._prefix+"/lib"):
                 os.makedirs(self._prefix+"/lib",0755)
-                os.symlink(self._prefix+"/lib",self._prefix+"/lib64")
+            os.symlink(self._prefix+"/lib",self._prefix+"/lib64")
 
         # Enviorment vars
         prefix = os.environ['PREFIX'] = self._prefix
